\name{pgsc-package}
\alias{pgsc-package}
\alias{pgsc}
\docType{package}
\title{
\packageTitle{pgsc}
}
\description{
\packageDescription{pgsc}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{pgsc}
\packageIndices{pgsc}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{pgsc}

Maintainer: \packageMaintainer{pgsc}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
