% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pgsc.dta}
\alias{pgsc.dta}
\title{Synthetic data for PGSC testing.}
\format{A data frame with 750 rows and 8 variables:
\describe{
  \item{n}{The unit, here labeled as a US state}
  \item{t}{The time period}
  \item{y}{The outcome variable}
  \item{D1}{The first treatment variable}
  \item{D2}{The second treatment variable}
  \item{X1}{The first observed confounding factor}
  \item{X2}{The second observed confounding factor}
  \item{X3}{The third observed confounding factor}
}}
\source{
Generated by code in the package vignette "Using \code{pgsc}".
}
\usage{
pgsc.dta
}
\description{
A dataset with an outcome given by a treatment and a set of factors.
}
\keyword{datasets}
