% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{W_extract}
\alias{W_extract}
\title{Extract weights}
\usage{
W_extract(wt, N)
}
\arguments{
\item{wt}{A N*(N-2) matrix of minimal weights}

\item{N}{The number of rows of the input matrix}
}
\value{
A matrix of full weights
}
\description{
Converts an N*(N-2) matrix of minimal weights into a full N*N matrix of weights.
The minimal weights are missing the diagonal entries (which are all zero in the output) 
plus the final column of weights. The full weights include explicit zeroes for the diagonal
and the final column such that each row adds up to one.
}
\examples{
wt <- matrix( c(.2, .2, .4, .3, .5, .4, .6, .1), nrow = 4, ncol=2 )
W_extract(wt, 4)

}

